[utils] Moved yaml node related utils functions into a separated translation unit and changed all the corresponding code.
